requirements.txt
captcha==0.3
click==7.1.2
Django==2.2
django-crontab==0.7.1
django-markdown==0.8.4
django-mdeditor==0.1.18
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
Flask==1.1.2
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==8.0.1
Pygments==2.7.3
PyMySQL==0.9.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.0
